refactor(query): extract tool-result budgeting into runner/tool_budget.rs
